Regroupe ce qui est commun à des redevances et substances
